[FEAT] 워크스페이스 설정 로직

+++ src/main/java/com/tiki/server/memberteammanager/service/MemberTeamManagerService.java
@@ -7,7 +7,7 @@ import com.tiki.server.memberteammanager.entity.MemberTeamManager;
 import com.tiki.server.note.adapter.NoteFinder;
 import com.tiki.server.note.entity.Note;
 import com.tiki.server.team.exception.TeamException;
-import com.tiki.server.memberteammanager.service.dto.response.MemberTeamPositionGetResponse;
+import com.tiki.server.memberteammanager.service.dto.response.MemberTeamInformGetResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,22 +31,24 @@ public class MemberTeamManagerService {
 
     @Transactional
     public void kickOutMemberFromTeam(final long memberId, final long teamId, final long kickOutMemberId) {
-        checkIsAdmin(memberId, teamId);
-        MemberTeamManager memberTeamManager = memberTeamManagerFinder.findByMemberIdAndTeamId(kickOutMemberId, teamId);
+        MemberTeamManager accessMember = memberTeamManagerFinder.findByMemberIdAndTeamId(memberId, teamId);
+        checkIsAdmin(accessMember);
+        MemberTeamManager kickOutMember = memberTeamManagerFinder.findByMemberIdAndTeamId(kickOutMemberId, teamId);
         deleteNoteDependency(kickOutMemberId, teamId);
-        memberTeamManagerDeleter.delete(memberTeamManager);
+        memberTeamManagerDeleter.delete(kickOutMember);
     }
 
     @Transactional
     public void leaveTeam(final long memberId, final long teamId) {
-        MemberTeamManager memberTeamManager = checkIsNotAdmin(memberId, teamId);
+        MemberTeamManager accessMember = memberTeamManagerFinder.findByMemberIdAndTeamId(memberId, teamId);
+        checkIsNotAdmin(accessMember);
         deleteNoteDependency(memberId, teamId);
-        memberTeamManagerDeleter.delete(memberTeamManager);
+        memberTeamManagerDeleter.delete(accessMember);
     }
 
-    public MemberTeamPositionGetResponse getPosition(final long memberId, final long teamId) {
+    public MemberTeamInformGetResponse getMemberTeamInform(final long memberId, final long teamId) {
         MemberTeamManager memberTeamManager = memberTeamManagerFinder.findByMemberIdAndTeamId(memberId, teamId);
-        return MemberTeamPositionGetResponse.from(memberTeamManager.getPosition());
+        return MemberTeamInformGetResponse.of(memberTeamManager.getPosition(), memberTeamManager.getName());
     }
 
     @Transactional
@@ -55,19 +57,16 @@ public class MemberTeamManagerService {
         memberTeamManager.updateName(name);
     }
 
-    private void checkIsAdmin(final long memberId, final long teamId) {
-        MemberTeamManager memberTeamManager = memberTeamManagerFinder.findByMemberIdAndTeamId(memberId, teamId);
+    private void checkIsAdmin(MemberTeamManager memberTeamManager) {
         if (!memberTeamManager.getPosition().equals(ADMIN)) {
             throw new TeamException(INVALID_AUTHORIZATION_DELETE);
         }
     }
 
-    private MemberTeamManager checkIsNotAdmin(final long memberId, final long teamId) {
-        MemberTeamManager memberTeamManager = memberTeamManagerFinder.findByMemberIdAndTeamId(memberId, teamId);
+    private void checkIsNotAdmin(MemberTeamManager memberTeamManager) {
         if (memberTeamManager.getPosition().equals(ADMIN)) {
             throw new TeamException(TOO_HIGH_AUTHORIZATION);
         }
-        return memberTeamManager;
     }
 
     private void deleteNoteDependency(final long memberId, final long teamId) {
